changed mention types
