+++ src/commands/leaderboard.ts
@@ -1,4 +1,4 @@
-import { ActionRowBuilder, ButtonBuilder, ButtonInteraction, ButtonStyle, ChatInputCommandInteraction, ComponentType, EmbedBuilder, Message } from 'discord.js';
+import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, ComponentType, EmbedBuilder, Message } from 'discord.js';
 import { OptionalCommandArguments } from '../commands';
 import { postLeaderboard } from '../misc/leaderboards';
 import { getAllowedMentions } from '../misc/misc';
