+++ src/misc/leaderboards.ts
@@ -1,22 +1,5 @@
 import { Leaderboard, SortedLeaderboard, leaderboards } from '../raiha';
 
-export async function postRank(id: string) {
-  const { Native, Raiha, Loserboard } = leaderboards;
-  const [native, raiha, loser] = [Native, Raiha, Loserboard].map(leaderboard => {
-    const value = leaderboard[id] ?? 0;
-    const rankValues = Object.values(leaderboard);
-    return {
-      value,
-      rank: value ? rankValues.reduce((a, b) => a + (b > value ? 1 : 0), 0) + 1 : null,
-      total: rankValues.length
-    }
-  })
-  return `Leaderboard ranking for <@${id}>:\n` +
-    `__**Native**__\n${native.rank ? `#${native.rank}/${native.total}` : 'Unranked'} with a count of ${native.value}.\n` +
-    `__**Raiha**__\n${raiha.rank ? `#${raiha.rank}/${raiha.total}` : 'Unranked'} with a count of ${raiha.value}.\n` +
-    `__**Loserboard**__\n${loser.rank ? `#${loser.rank}/${loser.total}` : 'Unranked'} with a count of ${loser.value}.`;
-}
-
 export async function postLeaderboard(): Promise<Array<{ text: string, footer: string }>> {
   const { Native, Raiha, Statistics: { Requests } } = leaderboards;
   const [native, raiha] = [getSortedLeaderboard(Native), getSortedLeaderboard(Raiha)]
@@ -45,7 +28,7 @@ function getText(board: SortedLeaderboard, pageLength: number): string[] {
   return results.length ? results : ["No results."];
 }
 
-export function getSortedLeaderboard(leaderboard: Leaderboard): SortedLeaderboard {
+function getSortedLeaderboard(leaderboard: Leaderboard): SortedLeaderboard {
   return Object.entries(leaderboard)
     .map(([user, value]) => ({ user, value }))
     .sort((a, b) => b.value - a.value);
